Update database-searching.md
putting all import calls at the top for the reasons noted in the pairwise alignment doc:

"""
Added loading all packages at the top. Here's the thinking. I can't think of another way to avoid someone jumping into the middle and having a problem with having all the functions up to that point loaded. This happened to me several times while running through, and then i had to search for where the import was that i needed. Loading everything at the top helps this, and it allows removing them from being mixed the main text where people are thinking about concepts and it may be unnecessarily distracting. There may be clear reasons not to do this that i'm not thinking of, but if not, i think this may be smoother.
"""

I think even if you want them in-place so people more easily know where things are coming from, it's still helpful to have a code block with all of them at the top.